Unit tests for PSP-3693 Add Notes (#1741)
* Add validation schema for adding notes

* Refactor add note forms

* documentation

* Unit tests for AddNotesForm

* Unit tests for AddNotesContainer

* Unit tests for API hooks

* Fix test warnings

* Unit tests - form state hook

* Use long for primary keys instead of int

* Backend controller unit tests

* Backend note service tests

* NoteRepository unit tests

* Test corrections

+++ backend/api/Areas/Notes/Controllers/NoteController.cs
@@ -92,7 +92,7 @@ namespace Pims.Api.Areas.Notes.Controllers
         [HasPermission(Permissions.NoteDelete)]
         [ProducesResponseType(typeof(bool), 200)]
         [SwaggerOperation(Tags = new[] { "note" })]
-        public IActionResult DeleteNote(NoteType type, int noteId)
+        public IActionResult DeleteNote(NoteType type, long noteId)
         {
             _noteService.DeleteNote(type, noteId);
             return new JsonResult(true);
